+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Some things may be added, some things may be removed, and some things may look d
 * Self-contained projects that contain all images inside the project directory no longer prompt the user to update URIs if moved (https://github.com/qupath/qupath/pull/1668)
 * Channel names can now be set for all fluorescence images, even if they are RGB (https://github.com/qupath/qupath/pull/1659)
   * Note that channel colors still only be set for non-RGB images
+* Improved ImageJ integration (https://github.com/qupath/qupath/pull/1676 https://github.com/qupath/qupath/pull/1677)
 
 ### Experimental features
 These features are included for testing and feedback.